Isolate recent crash handler
